mv msg and form handler to hooks.ts
